Ajuste no script de instalacao do puppet para remover a versao instalada pelo vagrant

+++ scripts/puppet_install.sh
@@ -1,21 +1,26 @@
 #!/bin/bash
 
-#set -e -x
+# Remove puppet e outras gemas ruby instaladas pelo vagrant
+if [ -d /opt/vagrant_ruby ]; then
+  echo "Removing directory /opt/vagrant_ruby from system and PATH"
+  PATH=${PATH#/opt/vagrant_ruby/bin:}
+  PATH=${PATH%:/opt/vagrant_ruby/bin}
+  rm -rf /opt/vagrant_ruby
+fi
 
-# Garantir que a versao suporte config de proxy
-#if which puppet > /dev/null : them
-#  echo "Puppet is already installed"
-#  exit 0
-#fi
+if which puppet >/dev/null ; then
+  echo "Puppet is already installed"
+  exit 0
+fi
 
 apt-get update
-wget -qO /tmp/puppetlabs-release-precise.deb https://apt.puppetlabs.com/puppetlabs-release-precise.deb
+wget -qO /tmp/puppetlabs-release-precise.deb \
+	https://apt.puppetlabs.com/puppetlabs-release-precise.deb
 dpkg -i /tmp/puppetlabs-release-precise.deb
 rm /tmp/puppetlabs-release-precise.deb
 apt-get update
 echo "Installing puppet"
 apt-get install --yes puppet
-#puppet resource package puppet ensure=latest
 echo "Puppet installed!"
 
 exit 0
